+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/events/LogMinerEventRow.java
@@ -128,7 +128,7 @@ public class LogMinerEventRow {
         return info;
     }
 
-    public int getSsn() {
+    public long getSsn() {
         return ssn;
     }
 
@@ -179,7 +179,7 @@ public class LogMinerEventRow {
         this.redoSql = getSqlRedo(resultSet);
         this.status = resultSet.getInt(STATUS);
         this.info = resultSet.getString(INFO);
-        this.ssn = resultSet.getInt(SSN);
+        this.ssn = resultSet.getLong(SSN);
         this.thread = resultSet.getInt(THREAD);
         if (this.tableName != null) {
             this.tableId = new TableId(catalogName, tablespaceName, tableName);
